Moved Gyroscopic Cockpit to B-Wing class. Code cleanup/standardization
